chore: move stories to the correct folder
